+++ Legacy Code/.gitignore
@@ -1,0 +1,12 @@
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+lerna-debug.log*
+node_modules/
+package-lock.json
+npm
+.eslintcache
+session.json
+.cache/
