+++ Legacy Code/README.md
@@ -1,0 +1,113 @@
+<!-- <h1 align="center">
+  <a href="https://github.com/TheWhatsBot/WhatsBot"><img src="https://telegra.ph/file/96ccad5945c18944c5f15.png" alt="whatsbot" width="290"></a>
+  <br>
+<b>Whatsbot</b>
+</h1>
+<h4 align="center">Moduler Userbot for Whatsapp</h4> -->
+
+![WhatsBot](https://socialify.git.ci/TheWhatsBot/WhatsBot/image?description=1&font=Source%20Code%20Pro&forks=1&language=1&owner=1&pattern=Floating%20Cogs&stargazers=1&theme=Dark) <br>
+
+<p align="center">
+<a href="https://github.com/TheWhatsBot/WhatsBot/releases">
+    <img src="https://shields.io/badge/WHATSBOT-Version--1.6.0-red?logo=whatsapp&style=for-the-badge"
+         alt="Version"></a><br>
+   <a href="https://github.com/TheWhatsBot/WhatsBot/wiki">
+ <img src="https://shields.io/badge/WIKI-red?style=for-the-badge" alt="Wiki"></a>
+  <a href="https://github.com/TheWhatsBot/WhatsBot/wiki/Demo">
+ <img src="https://shields.io/badge/DEMO-red?style=for-the-badge" alt="Demo"></a>
+    <a href="#deploy-">
+<img src="https://shields.io/badge/DEPLOY-red?style=for-the-badge" alt="Deploy"></a>
+    <a href="#credits-">
+<img src="https://shields.io/badge/Credits-red?style=for-the-badge" alt="Credits"></a>
+   <a href="#inspiration-">
+ <img src="https://shields.io/badge/Inspiration-red?style=for-the-badge" alt="Inspiration"></a>
+   <a href="#develop-">
+ <img src="https://shields.io/badge/Develop-red?style=for-the-badge" alt="Develop"></a><br>
+    <a href="#contributors-">
+<img src="https://shields.io/badge/Whatsbot-Contributors-red?style=for-the-badge" alt="Contributors"></a>
+</p>
+
+<br>
+
+### Features :
+
+- Modular
+- Pmpermit Included
+- Easily Deployable
+- Ban Free (Using Whatsapp Web)
+- Easy to Use
+- Directory Listing in Web Server
+- Terminal Included
+- Encrypted Traffic
+
+### Pre Added Plugins :
+
+- **pmpermit** - Permission for direct message
+- **term** - Terminal in Whatsapp
+- **yt** - Download Youtube video from link
+- **weather** - Lookup today's weather
+- **carbon** - Generate beautiful image from text
+- **sing** - Download a song by search
+- **qr** - Generate QR from text
+- **directlink** - Get direct link of photos
+- **tr** - Translate Text
+- **ud** - Urban Dictionary
+- **sticker** - Make sticker from Image
+- **crypto** - Fetch Latest Price of a CryptoCurrency
+- **watch** - Find out where to watch a Movie/Show
+- **shorten** - Get Short URL
+- **ocr** - Extract Text from Image
+
+[More modules]("https://github.com/TheWhatsBot/WhatsBot/wiki/Commands")
+
+### Deploy :
+
+[![Deploy with Heroku](https://www.herokucdn.com/deploy/button.svg "Deploy with Heroku")](https://heroku.com/deploy?template=https://github.com/TheWhatsBot/WhatsBot "Deploy with Heroku")<br>
+[![Deploy+on+Railway](https://railway.app/button.svg)](<https://railway.app/new/template?template=https://github.com/TheWhatsBot/WhatsBot&plugins=mongodb&envs=SESSION,PMPERMIT_ENABLED,PMPERMIT_MUTETIME,DEFAULT_TR_LANG,ENABLE_DELETE_ALERT,YT_DATA_API_KEY,OCR_SPACE_API_KEY,INFOSPACE_API_KEY&SESSIONDesc=Puppeteer+Session.+Ge+it+by+running+genToken.js&PMPERMIT_ENABLEDDesc=Enable+Pmpermit+write+true+or+false+only&PMPERMIT_ENABLEDDefault=true&PMPERMIT_MUTETIMEDesc=How+many+seconds+an+user+get+muted+if+he+is+spamming.+Default+is+30+Minutes+(1800+Secs)&PMPERMIT_MUTETIMEDefault=1800&YT_DATA_API_KEYDesc=Youtube+DATA+API+key+grab+it+from+cloud.google.com&DEFAULT_TR_LANGDesc=Default+Translation+Language&DEFAULT_TR_LANGDefault=en&ENABLE_DELETE_ALERTDesc=If+true+and+if+someone+delete+message+in+PM,+Bot+will+send+the+deleted+message+in+that+chat+(Exclude+Media)&ENABLE_DELETE_ALERTDefault=true&OCR_SPACE_API_KEYDesc=Get+it+from+https://ocr.space/OCRAPI&INFOSPACE_API_KEYDesc=Get+it+from+https://infospace.club>)<br>
+Read the [wiki](https://github.com/TheWhatsBot/WhatsBot/wiki/Deploy-with-Heroku) before Deploy it with Heroku<br><br>
+[Deploy in VPS](https://github.com/TheWhatsBot/WhatsBot/wiki/Deploy-in-VPS)
+
+### Commands :
+
+_Send <code>!help</code> in Anywhere. Or read this [Command's Wiki](https://github.com/TheWhatsBot/WhatsBot/wiki/Commands "Command's Wiki")_
+
+### Develop :
+
+_You can develop your own module for Whatsbot. Check out the [wiki](https://github.com/TheWhatsBot/WhatsBot/wiki/Development), It is very easy if you know Javascript._
+
+### How it works :
+
+- Open whatsapp web in server using puppeteer
+- Listen messages and take action.
+
+### Inspiration :
+
+- [Userge](https://github.com/UsergeTeam/Userge "Userge") - This Project takes Pmpermit insprition from Userge
+
+### Credits :
+
+- [Whatsapp Web JS](https://github.com/pedroslopez/whatsapp-web.js/ "Whatsapp Web JS") - This project is fully depended on this Library
+- [Carbon API](https://github.com/cyberboysumanjay/Carbon-API "Carbon API") - Thanks to [Sumanjay](https://github.com/cyberboysumanjay "Sumanjay") for his Carbon API
+- [Google Translate API](https://github.com/iamtraction/google-translate-api "Google Translate API") - Unofficial Google Translator Library
+- [Musicder](https://github.com/cachecleanerjeet/Musicder "Musicder") & [Jiosaavn API ](https://github.com/cachecleanerjeet/JiosaavnAPI "Jiosaavn API ") - For Jiosaavn Module
+- [savefrom.net](https://savefrom.net/ "savefrom[dot]net") - For Youtube Module
+- [Weather API](https://github.com/cachecleanerjeet/weather-api "Weather API") - For Weather Module
+- [Cribuzz API](https://github.com/mskian/cricket-api "Cribuzz API") - Thanks to [mskian](https://github.com/mskian) for his great Cribuzz API
+
+### Contributors :
+
+Whatsbot is fully depended on it's community, Feel free to open a PR if you added a new feature. Thanks to these Great Developers for contributing.
+
+<a href="https://github.com/thewhatsbot/whatsbot/graphs/contributors">
+  <img src="https://contrib.rocks/image?repo=thewhatsbot/whatsbot" />
+</a>
+
+### License & Copyright :
+
+- This Project is [Apache-2.0](https://github.com/TheWhatsBot/WhatsBot/blob/main/LICENSE) Licensed
+- Copyright 2021 by [Tuhin Kanti Pal](https://github.com/cachecleanerjeet)
+
+### Connect :
+
+- [Channel](https://telegram.dog/whatsbotnews)
+- [Support Group](https://telegram.dog/whatsbotsupport)
