+++ Legacy Code/_config.yml
@@ -1,0 +1,1 @@
+theme: jekyll-theme-minimal
