+++ Legacy Code/heroku.yml
@@ -1,0 +1,3 @@
+build:
+  docker:
+    web: Dockerfile
