+++ Legacy Code/modules/ocr.js
