+++ Legacy Code/modules/qr.js
@@ -1,0 +1,14 @@
+const qr = require('qr-image');
+
+async function qrgen(text) {
+
+    const data = ({
+        mimetype: "image/png",
+        data: await (qr.imageSync(text, { type: 'png' }).toString('base64')),
+        filename: text + ".png"
+    })
+    return data
+}
+module.exports = {
+    qrgen
+}
