+++ Legacy Code/modules/tempdata/qr.html
@@ -1,0 +1,64 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Whatsbot</title>
+    <style>
+        body {
+            height: 90vh;
+            min-height: 600px;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            flex-direction: column;
+            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+        }
+
+        .qr {
+            padding: 15px 20px;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+        }
+
+        .qr img {
+            width: 250px;
+            padding: 15px 20px;
+        }
+
+        h2,
+        p {
+            width: 100%;
+            max-width: 350px;
+            text-align: center;
+            padding-left: 15px;
+            padding-right: 15px;
+        }
+    </style>
+</head>
+<h2>Scan the <span style="color:green">QR.</span></h2>
+<div class="qr">
+    <img id="qr" src="/qr.svg" alt="QrCode">
+</div>
+<p style="padding-bottom: 30px;">If you scaned it successfully, please go back on your terminal and copy the <span
+        style="color:green">JSON</span>
+</p>
+
+<h1><a style="color:green; text-decoration: none; padding-bottom: 35px;" target="_blank"
+        href="https://github.com/TheWhatsBot/WhatsBot">WhatsBot</a></h1>
+
+<body>
+
+
+</body>
+
+<script>
+    setInterval(() => {
+        document.getElementById("qr").src = "/qr.svg?ts=" + new Date().getTime();
+    }, 4000);
+</script>
+
+</html>
