+++ Legacy Code/public/readme.txt
@@ -1,0 +1,1 @@
+this directory will be served publicly
